修复 UserTunnelWithDetailDto.flow 类型为 Double，避免读取时截断小数

+++ springboot-backend/src/main/java/com/admin/common/dto/UserTunnelWithDetailDto.java
@@ -31,7 +31,7 @@ public class UserTunnelWithDetailDto {
     /**
      * 流量限制
      */
-    private Integer flow;
+    private Double flow;
     
     /**
      * 转发数量限制
